Type Checking, Initial Node Test
Tried to run `tsc -w`, but it didn't work because it couldn't resolve `@hono/hono/jsx`, changing it to `@jsr/hono__hono/jsx` makes it work now.

https://github.com/microsoft/TypeScript/issues/58600

I'm going to see what it looks like on making this project work closer to plain ESM/Node/npm, maybe with the intent of getting it to work in frontend JS too, I'm not sure it's setup for that just yet.