Named Imports
Started working on adding the proper typings for things with the new edits to `types.ts`, and accessing things with the namespace gets a bit taxing, so I want to migrate to this setup before implementing that.

Definitely going to look into using better-named named exports for NBTify, they don't have to be so general as I have them now.